Quick Fix: changed target of Plugin eofnewline
The Plugin "eofnewline" was missing the required `-- lite-xl 1.16` at the start of the file.
I duplicated the original file and added the line. The new file is called `eofnewline-xl.lua`.
A PR was submitted and accepted from the original author.
This commit simply changes the target from `eofnewline.lua` to `eofnewline-xl.lua`.