Improve the texcompile plugin to work on windows
Borrow code from the rxi/console implementation to reliably
run commands on windows. Ideally should be based on the new reproc
library but this has to be done.

Problems of the current implementation:

- the output of the commands is ignored including warning and error messages
- the exit code of the command process is ignored

As a modification from the previous version of the plugin we no longer use
environment variables but we rely on "config" variables.